+++ src/components/search-modal/search-modal.js
@@ -92,23 +92,23 @@ class SearchModal extends PureComponent {
     this.setState(({ isModalVisible }) => ({
       isModalVisible: !isModalVisible,
     }));
-  }
+  };
 
   close = () => {
     this.setState({
       isModalVisible: false,
     });
-  }
+  };
 
   updateSearch = () => {
     this.close();
     this.updateFilter(this.state.query);
-  }
+  };
 
   resetSearch = () => {
     this.close();
     this.updateFilter({});
-  }
+  };
 
   toggleFilter = filterName => () => {
     const filterToBeToggled = filterName === 'access-type'
@@ -123,7 +123,7 @@ class SearchModal extends PureComponent {
       [filterToBeToggled]: !currentState[filterToBeToggled],
       [filterToBeDisabled]: false,
     }));
-  }
+  };
 
   handleSearchFieldChange = (searchfield) => {
     this.setState(({ query }) => ({
@@ -132,7 +132,7 @@ class SearchModal extends PureComponent {
         searchfield,
       }),
     }));
-  }
+  };
 
   handleSearchQueryChange = q => {
     this.setState(({ query }) => ({
@@ -141,7 +141,7 @@ class SearchModal extends PureComponent {
         q,
       },
     }));
-  }
+  };
 
   handleFilterChange = (sort, filter) => {
     this.setState(({ query }) => ({
@@ -152,7 +152,7 @@ class SearchModal extends PureComponent {
         q: query.q,
       }),
     }));
-  }
+  };
 
   handleStandaloneFilterChange = filter => {
     this.setState(({ query }) => ({
@@ -164,7 +164,7 @@ class SearchModal extends PureComponent {
       this.updateFilter(this.state.query);
       this.close();
     });
-  }
+  };
 
   render() {
     const {