+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change history for ui-eholdings
 
-## [10.1.0] (IN PROGRESS)
+## [11.0.0] (IN PROGRESS)
 
 * Add `target="_blank"` to the `New` button of agreements accordion. (UIEH-1446)
 * Remove `shouldFocus` HOC and all related code. (UIEH-1444)
-* Package Record - Title List Accordion - Improve Visibility of Search Within/Filter/Sort options. (UIEH-1440)
+* *BREAKING* Package Record - Title List Accordion - Improve Visibility of Search Within/Filter/Sort options. (UIEH-1440)
+* *BREAKING* Package Record - Title List view - Apply MCL component, add display of managed, custom coverage dates. (UIEH-1442)
 
 ## [10.0.1] (https://github.com/folio-org/ui-eholdings/tree/v10.0.1) (2024-11-12)